DOMA-1722 changed implementation of SBBOL credential management in favour of functions instead of a class
Uplifted initialising of context to top-level caller.
