+++ apps/condo/domains/organization/integrations/sbbol/utils/index.js
@@ -1,0 +1,11 @@
+const { buildBicryptId } = require('./buildBicryptId')
+const { changeClientSecret } = require('./changeClientSecret')
+const { getOrganizationAccessToken } = require('./getOrganizationAccessToken')
+const { refreshAllTokens } = require('./refreshAllTokens')
+
+module.exports = {
+    buildBicryptId,
+    changeClientSecret,
+    getOrganizationAccessToken,
+    refreshAllTokens,
+}
